// server/tests/schema.d.ts
export interface paths {
  '/api/v1/users': {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    get: {
      parameters: {
        query?: never
        header?: never
        path?: never
        cookie?: never
      }
      requestBody?: never
      responses: {
        /** @description List Users */
        200: {
          headers: {
            [name: string]: unknown
          }
          content: {
            'application/json': components['schemas']['User'][]
          }
        }
        /** @description Internal Server Error */
        500: {
          headers: {
            [name: string]: unknown
          }
          content: {
            'application/json': {
              message: string
            }
          }
        }
      }
    }
    put?: never
    post: {
      parameters: {
        query?: never
        header?: never
        path?: never
        cookie?: never
      }
      requestBody: {
        content: {
          'application/json': {
            /** @example John Doe */
            name: string
            /**
             * Format: email
             * @example john_doe@example.com
             */
            email: string
          }
        }
      }
      responses: {
        /** @description Create User */
        200: {
          headers: {
            [name: string]: unknown
          }
          content: {
            'application/json': {
              /**
               * Format: uuid
               * @example 778e7155-798c-4f06-ba52-e93535a60069
               */
              id: string
            }
          }
        }
        /** @description Bad Request Error */
        400: {
          headers: {
            [name: string]: unknown
          }
          content: {
            'application/json': {
              message: string
            }
          }
        }
        /** @description Already Exist Error */
        409: {
          headers: {
            [name: string]: unknown
          }
          content: {
            'application/json': {
              message: string
            }
          }
        }
        /** @description Internal Server Error */
        500: {
          headers: {
            [name: string]: unknown
          }
          content: {
            'application/json': {
              message: string
            }
          }
        }
      }
    }
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
  '/api/v1/users/{userId}': {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    get: {
      parameters: {
        query?: never
        header?: never
        path: {
          userId: string
        }
        cookie?: never
      }
      requestBody?: never
      responses: {
        /** @description Get User */
        200: {
          headers: {
            [name: string]: unknown
          }
          content: {
            'application/json': components['schemas']['User']
          }
        }
        /** @description Bad Request Error */
        400: {
          headers: {
            [name: string]: unknown
          }
          content: {
            'application/json': {
              message: string
            }
          }
        }
        /** @description Not Found Error */
        404: {
          headers: {
            [name: string]: unknown
          }
          content: {
            'application/json': {
              message: string
            }
          }
        }
        /** @description Internal Server Error */
        500: {
          headers: {
            [name: string]: unknown
          }
          content: {
            'application/json': {
              message: string
            }
          }
        }
      }
    }
    put: {
      parameters: {
        query?: never
        header?: never
        path: {
          userId: string
        }
        cookie?: never
      }
      requestBody: {
        content: {
          'application/json': {
            /** @example John Doe */
            name?: string
            /**
             * Format: email
             * @example john_doe@example.com
             */
            email?: string
          }
        }
      }
      responses: {
        /** @description Update Users */
        200: {
          headers: {
            [name: string]: unknown
          }
          content: {
            'application/json': Record<string, never>
          }
        }
        /** @description Bad Request Error */
        400: {
          headers: {
            [name: string]: unknown
          }
          content: {
            'application/json': {
              message: string
            }
          }
        }
        /** @description Not Found Error */
        404: {
          headers: {
            [name: string]: unknown
          }
          content: {
            'application/json': {
              message: string
            }
          }
        }
        /** @description Already Exist Error */
        409: {
          headers: {
            [name: string]: unknown
          }
          content: {
            'application/json': {
              message: string
            }
          }
        }
        /** @description Internal Server Error */
        500: {
          headers: {
            [name: string]: unknown
          }
          content: {
            'application/json': {
              message: string
            }
          }
        }
      }
    }
    post?: never
    delete: {
      parameters: {
        query?: never
        header?: never
        path: {
          userId: string
        }
        cookie?: never
      }
      requestBody?: never
      responses: {
        /** @description Delete User */
        200: {
          headers: {
            [name: string]: unknown
          }
          content: {
            'application/json': Record<string, never>
          }
        }
        /** @description Bad Request Error */
        400: {
          headers: {
            [name: string]: unknown
          }
          content: {
            'application/json': {
              message: string
            }
          }
        }
        /** @description Not Found Error */
        404: {
          headers: {
            [name: string]: unknown
          }
          content: {
            'application/json': {
              message: string
            }
          }
        }
        /** @description Internal Server Error */
        500: {
          headers: {
            [name: string]: unknown
          }
          content: {
            'application/json': {
              message: string
            }
          }
        }
      }
    }
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
}
export type webhooks = Record<string, never>
export interface components {
  schemas: {
    User: {
      /**
       * Format: uuid
       * @example 778e7155-798c-4f06-ba52-e93535a60069
       */
      id: string
      /** @example John Doe */
      name: string
      /**
       * Format: email
       * @example john_doe@example.com
       */
      email: string
      /** @example 2024-11-04T10:59:02.136Z */
      createdAt: string
      /** @example 2024-11-04T10:59:02.136Z */
      updatedAt: string
    }
  }
  responses: never
  parameters: never
  requestBodies: never
  headers: never
  pathItems: never
}